requirements.txt
networkx==3.3
matplotlib==3.9.2
scipy==1.14.1
python-louvain==0.16
pyvis==0.3.2
nltk
wordcloud
sentence-transformers
seaborn
umap-learn
pypinyin
pyLDAvis